+++ terraform/ces-module/main.tf
@@ -129,5 +129,7 @@ module "ces-blueprint" {
 
   dogus = var.dogus
 
+  blueprint = var.blueprint
+
 }