+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [v5.7.0] - 2026-02-04
+### Added
+- [#125] Add snapshot crd required by the backup stack.
+- [#129] Add optional Blueprint parameter for the ces-blueprint terraform module.
+
 ## [v5.6.1] - 2025-11-26
 ### Fixed
 - [#123] fix values template of the ces-core terraform module